fix: herramientas mediofuncional (falta manuales, falta imagenes)
